fix build on Windows
